# docs/openapi/catalogo.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Catálogo de filmes — API pública",
    "version": "1.0.0",
    "description": "Sessão pelo cookie HttpOnly token. Faça login em /login.html no mesmo navegador ou execute POST /api/auth/login com uma conta local. Não cole o cookie no Authorize. GET /api/auth/me sem sessão retorna 401 e é uma chamada segura para demonstrar Try it out. Papel comum: usuario; administrador: admin. Alterações de dados devem ser testadas somente no ambiente local."
  },
  "servers": [
    {
      "url": "/",
      "description": "Mesma origem do catálogo (local ou publicada)"
    }
  ],
  "components": {
    "schemas": {
      "Mensagem": {
        "type": "object",
        "properties": {
          "mensagem": {
            "type": "string"
          }
        },
        "required": [
          "mensagem"
        ]
      },
...
      "Perfil": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "example": 1
          },
          "nome": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "foto_chave": {
            "type": "string",
            "nullable": true
          },
          "foto_url": {
            "type": "string",
            "nullable": true
          },
          "favoritos": {
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            }
          },
          "proprio": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "nome",
          "bio",
          "foto_chave",
          "foto_url",
          "favoritos",
          "proprio"
        ]
      },
...
    },
    "securitySchemes": {
      "sessao": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "Cookie HttpOnly criado por POST /api/auth/login. Faça login pelo formulário /login.html no mesmo navegador ou execute o login no Swagger. O navegador envia o cookie automaticamente; não leia nem cole o cookie no Authorize. SameSite=Lax; Secure em produção; duração 8 horas."
      }
    }
  },
  "paths": {
    "/api/perfil/fotos/{usuarioId}/{arquivo}": {
      "get": {
        "tags": [
          "Perfil e upload"
        ],
        "summary": "Ler foto publicamente",
        "description": "Sem autenticação. Só aceita chave perfis/ID/UUID.(jpg|png|webp). Leitura pública via catálogo; escrita restrita ao backend.",
        "operationId": "get_api_perfil_fotos_usuarioId_arquivo",
        "security": [],
        "responses": {
          "200": {
            "description": "Foto; Cache-Control public, max-age=86400 e X-Content-Type-Options nosniff.",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Chave inválida ou foto não encontrada",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Not Found"
                }
              }
            }
          },
          "503": {
            "description": "Armazenamento indisponível ou resposta inválida",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Service Unavailable"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "usuarioId",
            "in": "path",
            "required": true,
            "description": "Dígitos decimais positivos, sem zero à esquerda (validado no caminho da chave).",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            }
          },
          {
            "name": "arquivo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f-]{36}\\.(jpg|png|webp)$",
              "example": "00000000-0000-0000-0000-000000000000.jpg"
            }
          }
        ]
      }
    },
    "/api/perfil/me": {
      "get": {
        "tags": [
          "Perfil e upload"
        ],
        "summary": "Consultar perfil",
        "description": "Perfil próprio (me) ou de outro usuário, incluindo IDs dos favoritos. Não exige propriedade para leitura. Cache-Control no-store. Requer sessão autenticada (cookie HttpOnly token).",
        "operationId": "get_api_perfil_me",
        "security": [
          {
            "sessao": []
          }
        ],
        "responses": {
          "200": {
            "description": "Perfil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "perfil": {
                      "$ref": "#/components/schemas/Perfil"
                    }
                  },
                  "required": [
                    "perfil"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Usuário não autenticado. / Token não informado. / Sessão inválida ou expirada. / Token com usuário inválido. / Usuário não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Usuário não autenticado."
                    }
                  },
                  "1": {
                    "value": {
                      "mensagem": "Token não informado."
                    }
                  },
                  "2": {
                    "value": {
                      "mensagem": "Sessão inválida ou expirada."
                    }
                  },
                  "3": {
                    "value": {
                      "mensagem": "Token com usuário inválido."
                    }
                  },
                  "4": {
                    "value": {
                      "mensagem": "Usuário não encontrado."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Perfil não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Perfil não encontrado."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Erro interno do servidor."
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Não foi possível consultar o perfil. / Serviço de autenticação indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mensagem"
                },
                "examples": {
                  "0": {
                    "value": {
                      "mensagem": "Não foi possível consultar o perfil."
                    }
                  },
                  "1": {
                    "value": {
                      "mensagem": "Serviço de autenticação indisponível."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}